+++ core/infrastructure/image.go
@@ -29,41 +29,23 @@ import (
 const (
 	annotationKeyImageTriggers         = "image.openshift.io/triggers"
 	annotationValueImageTriggersFormat = "[{\"from\":{\"kind\":\"ImageStreamTag\",\"name\":\"%s\"},\"fieldPath\":\"spec.template.spec.containers[?(@.name==\\\"%s\\\")].image\"}]"
-	// LatestTag the default name for latest image tag
-	LatestTag        = "latest"
+
 	versionSeparator = "."
-)
-
-const (
-
+	// LatestTag the default name for latest image tag
+	LatestTag = "latest"
 	// DefaultImageRegistry the default services image repository
 	DefaultImageRegistry = "quay.io"
 	// DefaultImageNamespace the default services image namespace
 	DefaultImageNamespace = "kiegroup"
-	// KogitoQuarkusUbi8Image Quarkus runtime builder image
-	KogitoQuarkusUbi8Image = "kogito-quarkus-ubi8"
-	// KogitoQuarkusJVMUbi8Image Quarkus jvm runtime builder image
-	KogitoQuarkusJVMUbi8Image = "kogito-quarkus-jvm-ubi8"
-	// KogitoQuarkusUbi8s2iImage Quarkus s2i builder image
-	KogitoQuarkusUbi8s2iImage = "kogito-quarkus-ubi8-s2i"
-	// KogitoSpringBootUbi8Image SpringBoot runtime builder image
-	KogitoSpringBootUbi8Image = "kogito-springboot-ubi8"
-	// KogitoSpringBootUbi8s2iImage SpringBoot s2i builder image
-	KogitoSpringBootUbi8s2iImage = "kogito-springboot-ubi8-s2i"
-)
 
-var (
-	// KogitoImages maps the default Kogito Images on a matrix of RuntimeType and its purpose
-	KogitoImages = map[api.RuntimeType]map[bool]string{
-		api.QuarkusRuntimeType: {
-			true:  KogitoQuarkusUbi8s2iImage,
-			false: KogitoQuarkusJVMUbi8Image,
-		},
-		api.SpringBootRuntimeType: {
-			true:  KogitoSpringBootUbi8s2iImage,
-			false: KogitoSpringBootUbi8Image,
-		},
-	}
+	// KogitoBuilderImage Builder Image for Kogito
+	KogitoBuilderImage = "kogito-builder"
+	// KogitoRuntimeJVM Runtime Image for Kogito with  JRE
+	KogitoRuntimeJVM = "kogito-runtime-jvm"
+	//KogitoRuntimeNative Runtime Image for Kogito for Native Quarkus Application
+	KogitoRuntimeNative = "Kogito-runtime-native"
+	//RuntimeTypeKey Env key to switch between the runtime
+	RuntimeTypeKey = "RUNTIME_TYPE"
 )
 
 // ImageHandler describes the handler structure to handle Kogito Services Images