+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 ## Enhancements  
 - [KOGITO-4245](https://issues.redhat.com/browse/KOGITO-4245) - Use Infinispan CR spec data instead of status to fetch credentials Secret 
+- [KOGITO-2317](https://issues.redhat.com/browse/KOGITO-2317) - Unify Kogito Quarkus and Spring Images
 - [KOGITO-4059](https://issues.redhat.com/browse/KOGITO-4059) Refactor Kogito operator into internal and core component
 
 ## Bug Fixes